refactored argument parser
